+++ test/printer_graph_html_test.rb
@@ -51,7 +51,7 @@ class PrinterGraphHtmlTest < TestCase
     sort_method_with_column_number = {:total_time => 3, :self_time => 4, :wait_time => 5, :children_time => 6}
 
     sort_method_with_column_number.each_pair do |sort_method, n|
-      printer.print(output = '', :sort_method => sort_method)
+      printer.print(output = '', sort_method: sort_method)
       times = graph_html_output_nth_column_values(output, n)
       assert_sorted times
     end